Create black_jack_game
I just wanted to train my studies with IF, WHILE, FOR and SWITCH conditions/loops. I thought it would be easy, but I struggled a little bit. Hopefully I'll get better at this. I still don't quite understand why I had to "reverse" the "winner" in my last IF condition if the winner is actually who gets 21 or closer to it.

Any comment it's very much appreciated. ;)