JS Task List Course Challenge
This a JavaScript challenge to create a script that shows a Task List. The main task was to create a list on stored in the "localStorage" and to show it. Well, I added some other things too in order to practice what I currently have been learning on these past few days. Your comments are very well appreciated. Thanks!